Fixing Inheritance and Type Cast issues
1st issue
 I was using new instead of override for the adapter.
 This means that if used as the adapter class it's method
 would be called. However, if the base class was used it's
 method would be called. Turns out I needed to make the getRadius
 method of the RoundPeg class virtual and have the adapter override.

2nd issue
 I was using Int through all of my width and radius calculations,
 then when getting the distance to the corner of the square in the
 SquarePegClassAdapter I was type casting essentially cutting off
 resolution. Causing the large square peg to not fail the
 RoundHole.fits using the adapter. I resolved this by switching
 to double everywhere and removing the int type cast on the
 Math.Sqrt

+++ Adapter/RoundHole.cs
@@ -6,14 +6,14 @@ namespace Adapter
 {
     public class RoundHole
     {
-        private int m_radius;
+        private double m_radius;
 
-        public RoundHole(int radius)
+        public RoundHole(double radius)
         {
             m_radius = radius;
         }
 
-        public int getRadius()
+        public double getRadius()
         {
             return m_radius;
         }
